pyCODER/requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.3
Babel==2.8.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
django-appconf==1.0.4
django-countries==6.1.2
django-geoip2-extras==1.2
django-hitcount==1.3.2
django-ipware==3.0.0
django-phonenumber-field==4.0.0
django-seo==0.3.5
django-social-share==1.4.0
django-static-jquery==2.1.4
django-suit==0.2.25
django-summernote==0.8.11.6
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
doc8==0.8.0
docutils==0.16
geoip2==3.0.0
gunicorn==20.0.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
maxminddb==1.5.2
pbr==5.4.4
phonenumberslite==8.12.6
Pillow==7.0.0
psycopg2-binary==2.8.5
pycparser==2.19
pylibmc==1.6.1
python-dateutil==2.8.1
python-env-utils==0.4.1
python-memcached==1.59
pytz==2019.3
pywatchman==1.4.1
PyYAML==5.3
requests==2.22.0
restructuredtext-lint==1.3.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
stevedore==1.31.0
tblib==1.6.0
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
